Fix error where current year data was always marked as 'fixed'
For performance reasons, we cache the current queries.
For year data, we consider the cache "temporary" if the requested
data is still for this year, meaning that we should regenerate it
periodically.

The problem was that it didn't look at what year was requested but only
whether it was older than the current month -- maybe it was copied from
MonthCacheDescriptor?